Consensus: MOVEONLY: Move CheckTx and CheckBlock from main to consensus
